PcInterface taken out, comms working
Former-commit-id: e524b5877998f15cf1a54fa222efaf43b83d2756 [formerly 2289a48b5a63f20ddde3b7500c627b9dc5b8d8ce]
Former-commit-id: 6033c28162ac12e16eb268de88749c38aeb935e6

+++ Common/app/rx_helper.cpp
@@ -169,7 +169,7 @@ void receiveDataBuffer(void) {
 
 /* XXX: temp function while eventHandler gets integrated*/
 void copyIntoBuffRx(uint8_t *copyFrom) {
-    for (int i = 0; i < sizeof(buffRx); i++) {
+    for (size_t i = 0; i < sizeof(buffRx); i++) {
         buffRx[i] = copyFrom[i];
     }
 }
